mcp_server/requirements.txt
# AutomationEdge MCP Server dependencies
mcp[cli]>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
